ml/requirements.txt
# VaultVision ML Pipeline - Dependencies
# Vault Sync AI LLC

# Computer Vision & ML
ultralytics>=8.1.0
opencv-python>=4.9.0
numpy>=1.26.0

# Object Tracking
lap>=0.4.0

# Vector Store & Search
chromadb>=0.4.0

# LLM for Natural Language Query
anthropic>=0.40.0

# Task Queue (shared with backend)
celery[redis]>=5.3.0
redis>=5.0.0

# Utilities
Pillow>=10.0.0
tqdm>=4.66.0
pydantic>=2.0.0
python-dotenv>=1.0.0
uuid6>=2024.1.12
